+++ docs/source/Changelog.rst
@@ -27,6 +27,10 @@ Added
 -  Some limited support for the VARIANT and VARIANT_DESC variables found
    in KiCad 10. They can be used for things like 3D and print_pcb even
    for KiCad 6.
+-  Variants:
+
+   -  pre_transform filter to KiCad variants
+
 -  E/DRC:
 
    -  Dark mode support and icon to select it.